backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.41.0,<1.0.0
psycopg2-binary>=2.9.10,<3.0.0
asyncpg>=0.30.0,<1.0.0
paramiko>=3.5.0,<4.0.0
cryptography>=46.0.0,<47.0.0
bcrypt>=4.2.0,<5.0.0
pyjwt>=2.10.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.10.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
slowapi>=0.1.9,<1.0.0
